move popup and message to statistic folder
