+++ Src/MoneyFox.Ui/Views/Popups/SelectDateRangeDialogViewModel.cs
@@ -4,6 +4,7 @@ using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
 using Core.Common.Extensions;
 using Messages;
+using Statistics;
 
 internal sealed class SelectDateRangeDialogViewModel : BasePageViewModel
 {
